requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
pymongo>=4.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
Pillow>=10.0.0
